Make NetworkUtils less fat
Breaks dependency cycles
