+++ core/src/main/java/de/danoeh/antennapod/core/util/download/MediaSizeLoader.java
@@ -1,0 +1,78 @@
+package de.danoeh.antennapod.core.util.download;
+
+import android.text.TextUtils;
+import de.danoeh.antennapod.core.service.download.AntennapodHttpClient;
+import de.danoeh.antennapod.core.storage.DBWriter;
+import de.danoeh.antennapod.core.util.NetworkUtils;
+import de.danoeh.antennapod.model.feed.FeedMedia;
+import io.reactivex.Single;
+import io.reactivex.SingleOnSubscribe;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.schedulers.Schedulers;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import android.util.Log;
+import okhttp3.Response;
+
+import java.io.File;
+import java.io.IOException;
+
+public abstract class MediaSizeLoader {
+    private static final String TAG = "MediaSizeLoader";
+
+    public static Single<Long> getFeedMediaSizeObservable(FeedMedia media) {
+        return Single.create((SingleOnSubscribe<Long>) emitter -> {
+            if (!NetworkUtils.isEpisodeHeadDownloadAllowed()) {
+                emitter.onSuccess(0L);
+                return;
+            }
+            long size = Integer.MIN_VALUE;
+            if (media.isDownloaded()) {
+                File mediaFile = new File(media.getLocalMediaUrl());
+                if (mediaFile.exists()) {
+                    size = mediaFile.length();
+                }
+            } else if (!media.checkedOnSizeButUnknown()) {
+                // only query the network if we haven't already checked
+
+                String url = media.getDownload_url();
+                if (TextUtils.isEmpty(url)) {
+                    emitter.onSuccess(0L);
+                    return;
+                }
+
+                OkHttpClient client = AntennapodHttpClient.getHttpClient();
+                Request.Builder httpReq = new Request.Builder()
+                        .url(url)
+                        .header("Accept-Encoding", "identity")
+                        .head();
+                try {
+                    Response response = client.newCall(httpReq.build()).execute();
+                    if (response.isSuccessful()) {
+                        String contentLength = response.header("Content-Length");
+                        try {
+                            size = Integer.parseInt(contentLength);
+                        } catch (NumberFormatException e) {
+                            Log.e(TAG, Log.getStackTraceString(e));
+                        }
+                    }
+                } catch (IOException e) {
+                    emitter.onSuccess(0L);
+                    Log.e(TAG, Log.getStackTraceString(e));
+                    return; // better luck next time
+                }
+            }
+            Log.d(TAG, "new size: " + size);
+            if (size <= 0) {
+                // they didn't tell us the size, but we don't want to keep querying on it
+                media.setCheckedOnSizeButUnknown();
+            } else {
+                media.setSize(size);
+            }
+            emitter.onSuccess(size);
+            DBWriter.setFeedMedia(media);
+        })
+        .subscribeOn(Schedulers.io())
+        .observeOn(AndroidSchedulers.mainThread());
+    }
+}
